+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+# [2.0.0-rc.1]
+
+## Alpha Changes
+
+- Restored Navigraph tab in Layers quick settings
+- Changed Map Settings icon
+- Controller presence will now show how many aircraft are controlled
+- Smooth aircraft will now be disabled on high zoom levels
+- Fixed bookings showing when filter is active
+
+## Features and Improvements
+
+- Added a setting to move favorites to top right corner of screen on desktop - like on mobile version
+- Added a setting to pin favorites
+- Added point bearing distance support
+- Added ICAO in route field support
+- Added pilots section into each atc showing aircraft under control
+- Added a setting to highlight sector where you are currently flying
+
+## Bug Fixes
+
+- Fixed concorde NAT tracks parsing incorrectly
+- Fixed some incorrect airways parsing bugs
+
 # [2.0.0-beta.6]
 
 ## Alpha Changes